{"schema":"decodex/commit/1","summary":"Split state SQLite backend module","authority":"XY-1115"}
